interview_agent/requirements.txt
fastapi
uvicorn
pydantic
langchain
langchain-openai
langchain-community
langchain-google-genai
python-dotenv
redis
pandas
fpdf
